Report generator

See merge request rootteam/rootcause!117

+++ core/src/main/java/fr/uge/modules/api/model/linking/Computation.java
@@ -1,6 +1,4 @@
 package fr.uge.modules.api.model.linking;
 
-import java.math.BigDecimal;
-
-public record Computation(String token_type, String value_log_first, String value_log_second, BigDecimal proximity) {
+public record Computation(String token_type, String value_log_first, String value_log_second, Double proximity) {
 }
